⚡️ :: string으로 변경

+++ Hexagonal Architecture/hexagonal-application/src/main/kotlin/com/example/domain/feed/usecase/CreateFeedUseccase.kt
@@ -12,7 +12,7 @@ class CreateFeedUseccase(
     private val userSecurityPort: UserSecurityPort
 ) {
     fun excute(createFeedRequest: CreateFeedRequest) {
-        val userId = userSecurityPort.getCurrentUserId().toLong()
+        val userId = userSecurityPort.getCurrentUserId()
 
         val feed = Feed(
             id = null,
